+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## main
 
+- Colourise json/ruby output
+
 ## v0.3.0
 
 - Rename project to `rj`